fix [BishopMoveValidator] lossy conversion of int into byte (while performing arthimetic operation)

+++ TMChessServer/src/main/java/com/thinking/machines/chess/server/validators/BishopMoveValidator.java
@@ -7,10 +7,10 @@ public static byte validateMove(byte startRowIndex,byte startColumnIndex,byte de
 //restricting vertical and horizontal movement of the (bishop)
 if((startRowIndex==destinationRowIndex && startColumnIndex!=destinationColumnIndex) || (startRowIndex!=destinationRowIndex && startColumnIndex==destinationColumnIndex)) return 0; 
 //validating movement of bishop diagonally on the same tile
-byte d1=startRowIndex-destinationRowIndex;
-byte d2=startColumnIndex-destinationColumnIndex;
-if(d1<0) d1=d1*(-1);
-if(d2<0) d2=d2*(-1);
+byte d1=(byte) (startRowIndex-destinationRowIndex);
+byte d2=(byte) (startColumnIndex-destinationColumnIndex);
+if(d1<0) d1=(byte)(d1*(-1));
+if(d2<0) d2=(byte)(d2*(-1));
 if(d1!=d2)
 {
 return 0;
@@ -20,7 +20,7 @@ byte tile;
 byte e,f;
 if(destinationRowIndex<startRowIndex && destinationColumnIndex<startColumnIndex)
 {
-for(e=startRowIndex-1,f=startColumnIndex-1;e>destinationRowIndex;e--,f--)
+for(e=(byte)(startRowIndex-1),f=(byte)(startColumnIndex-1);e>destinationRowIndex;e--,f--)
 {
 tile=board[e][f];
 if(tile!=0) 
@@ -32,7 +32,7 @@ return 0;
 }else //path blocker for top-right
 if(destinationRowIndex<startRowIndex && startColumnIndex<destinationColumnIndex)
 {
-for(e=startRowIndex-1,f=startColumnIndex+1;e>destinationRowIndex;e--,f++)
+for(e=(byte)(startRowIndex-1),f=(byte)(startColumnIndex+1);e>destinationRowIndex;e--,f++)
 {
 tile=board[e][f];
 if(tile!=0)
@@ -44,7 +44,7 @@ return 0;
 }else//path blocker for bottom-left
 if(startRowIndex<destinationRowIndex && destinationColumnIndex<startColumnIndex)
 {
-for(e=startRowIndex+1,f=startColumnIndex-1;e<destinationRowIndex;e++,f--)
+for(e=(byte)(startRowIndex+1),f=(byte)(startColumnIndex-1);e<destinationRowIndex;e++,f--)
 {
 tile=board[e][f];
 if(tile!=0) 
@@ -56,7 +56,7 @@ return 0;
 }else//path blocker for bottom-right
 if(startRowIndex<destinationRowIndex && startColumnIndex<destinationColumnIndex)
 {
-for(e=startRowIndex+1,f=startColumnIndex+1;e<destinationRowIndex;e++,f++)
+for(e=(byte)(startRowIndex+1),f=(byte)(startColumnIndex+1);e<destinationRowIndex;e++,f++)
 {
 tile=board[e][f];
 if(tile!=0)
